+++ vitest.config.ts
@@ -4,7 +4,12 @@ export default defineConfig({
   test: {
     coverage: {
       include: ["src/**/*.ts"],
-      exclude: ["src/**/*.test.ts", "src/types.ts"],
+      exclude: [
+        "src/**/*.test.ts",
+        "src/types.ts",
+        "src/middleware.ts",
+        "src/index.ts",
+      ],
     },
   },
 });